Revert "Revert "Finished - 59 make lyrics component independent""
