+++ src/g-player-react/src/Components/player/Player.js
@@ -155,6 +155,8 @@ export const Player = () => {
                 break;
         }
 
+        if(library.length === 0)library = tracks;
+
         if(isShuffle && library.length>0){
             library = getShuffledTrackList(library);
         }
